Fix Stepper component to use StepperContext
Update TypeScript version to 5.3.2

Update eslint to version 8.54.0

Refactor Stepper.component.tsx

test type

test

Remove redundant code in Stepper.component.tsx
