+++ package.json
@@ -66,11 +66,13 @@
     "@types/react-window": "^1.8.5",
     "@types/styled-components": "^5.1.25",
     "@types/styled-system": "^5.1.15",
+    "@typescript-eslint/eslint-plugin": "^6.12.0",
+    "@typescript-eslint/parser": "^6.12.0",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
     "csstype": "^3.1.0",
     "downshift": "^7.0.5",
-    "eslint": "^8.15.0",
+    "eslint": "^8.54.0",
     "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-storybook": "^0.6.15",
     "husky": "^3.0.5",
@@ -98,7 +100,7 @@
     "storybook": "^7.5.3",
     "styled-components": "^5.2.1",
     "styled-system": "^5.1.5",
-    "typescript": "^4.6.4",
+    "typescript": "^5.3.2",
     "vega": "^5.17.3",
     "vega-embed": "^6.0.0",
     "vega-lite": "^5.0.0",
